requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
PyYAML>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
gradio>=4.0.0
pytest>=7.0.0
